Moved Whitelabel tests to a directory
